[v7]: Re-enable diff tests
These tests now require `Configure` which requires AWS creds.

fixes #5518
